+++ code/basic/project3/src/matrix.c
@@ -3,7 +3,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-int createMatrix(struct Matrix* Mat, int rows, int cols){
+int createMatrix(struct Matrix* Mat, size_t rows, size_t cols){
     // 输入判断
     if(rows <= 0 || cols <= 0) return -1;
 
@@ -27,11 +27,11 @@ int deleteMatrix(struct Matrix* Mat){
 
 int copyMatrix(struct Matrix *destMat, const struct Matrix *srcMat){
     if((destMat->ROW_NUM != srcMat->ROW_NUM) || (destMat->COL_NUM != srcMat->COL_NUM)) return -1;
-    int rows = destMat->ROW_NUM;
-    int cols = destMat->COL_NUM;
+    size_t rows = destMat->ROW_NUM;
+    size_t cols = destMat->COL_NUM;
 
-    for(int i = 0; i < rows; i++){
-        for(int j = 0; j < cols; j++)
+    for(size_t i = 0; i < rows; i++){
+        for(size_t j = 0; j < cols; j++)
         destMat->data[i*cols + j] = srcMat->data[i*cols + j];
     }
 
@@ -40,11 +40,11 @@ int copyMatrix(struct Matrix *destMat, const struct Matrix *srcMat){
 
 int addMatrix(struct Matrix *MatA, struct Matrix *MatB){
     if((MatA->ROW_NUM != MatB->ROW_NUM) || (MatA->COL_NUM != MatB->COL_NUM)) return -1;
-    int rows = MatA->ROW_NUM;
-    int cols = MatA->COL_NUM;
+    size_t rows = MatA->ROW_NUM;
+    size_t cols = MatA->COL_NUM;
 
-    for(int i = 0; i < rows; i++){
-        for(int j = 0; j < cols; j++)
+    for(size_t i = 0; i < rows; i++){
+        for(size_t j = 0; j < cols; j++)
         MatA->data[i*cols + j] += MatB->data[i*cols + j];
     }
 
@@ -57,15 +57,15 @@ int addMatrix(struct Matrix *MatA, struct Matrix *MatB){
 int matmulMatrix(const struct Matrix *MatA, const struct Matrix *MatB, struct Matrix *MatC){
     // check
     if(MatA->ROW_NUM != MatC->ROW_NUM || MatA->COL_NUM != MatB->ROW_NUM || MatB->COL_NUM != MatC->ROW_NUM) return -1;
-    int rowA = MatA->ROW_NUM;
-    int colA = MatA->COL_NUM;
-    int colB = MatB->COL_NUM;
+    size_t rowA = MatA->ROW_NUM;
+    size_t colA = MatA->COL_NUM;
+    size_t colB = MatB->COL_NUM;
 
     // cal
-    for(int i = 0; i < rowA; i++){
-        for(int j = 0; j < colB; j++){
+    for(size_t i = 0; i < rowA; i++){
+        for(size_t j = 0; j < colB; j++){
             float temp = 0.0f;
-            for(int k = 0; k < colA; k++){
+            for(size_t k = 0; k < colA; k++){
                 temp += MatA->data[i*colA + k] * MatB->data[k*colB + j];
             }
             MatC->data[i*colB + j] = temp;
@@ -75,7 +75,7 @@ int matmulMatrix(const struct Matrix *MatA, const struct Matrix *MatB, struct Ma
 }
 
 
-int setMatrixValue(struct Matrix *Mat, int row, int col, float value){
+int setMatrixValue(struct Matrix *Mat, size_t row, size_t col, float value){
     if(row > Mat->ROW_NUM || col > Mat->COL_NUM) return -1;
     Mat->data[row*Mat->COL_NUM + col] = value;
     return 0;
@@ -83,13 +83,13 @@ int setMatrixValue(struct Matrix *Mat, int row, int col, float value){
 
 int displayMatrix(const struct Matrix *Mat){
     if(!Mat->data) return -1;
-    int row = Mat->ROW_NUM;
-    int col = Mat->COL_NUM;
+    size_t row = Mat->ROW_NUM;
+    size_t col = Mat->COL_NUM;
 
     printf("[");
-    for(int i = 0; i < row; i++){
+    for(size_t i = 0; i < row; i++){
         printf("[");
-        for(int j = 0; j < col; j++){
+        for(size_t j = 0; j < col; j++){
             if(j != col-1 ) printf("%.4f, ", Mat->data[i*col + j]);
             else printf("%.4f", Mat->data[i*col + j]);
         }
